+++ CarProjectMVC/CarProjectMVC.csproj
@@ -31,7 +31,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Context\" />
     <Folder Include="Views\Update\" />
   </ItemGroup>
