+++ CarProjectMVC/JWT/JwtToken.cs
@@ -1,9 +1,18 @@
 ﻿namespace CarProjectMVC.JWT
 {
+    /// <summary>
+    /// JWT токен для аутентификации
+    /// </summary>
     public class JwtToken
     {
+        /// <summary>
+        /// Токен для доступа
+        /// </summary>
         public string? AccessToken { get; set; }
 
+        /// <summary>
+        /// Токен для обновления
+        /// </summary>
         public string? RefreshToken { get; set; }
     }
 }
